agent-services/langgraph-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langgraph==0.0.40
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-core>=0.1.46,<0.2.0
pydantic>=2.5.0
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0
